bloc listener movement and pause splash screen
