+++ lib/credentials/list/view/credentials_list_page.dart
@@ -1,14 +1,7 @@
-import 'package:dio/dio.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:talao/app/interop/issuer/check_issuer.dart';
-import 'package:talao/app/interop/issuer/models/issuer.dart';
-import 'package:talao/app/interop/network/network_client.dart';
-import 'package:talao/app/shared/constants.dart';
-import 'package:talao/app/shared/error_handler/error_handler.dart';
 import 'package:talao/app/shared/model/credential_model/credential_model.dart';
-import 'package:talao/app/shared/widget/confirm_dialog.dart';
 import 'package:talao/credentials/widget/list_item.dart';
 import 'package:talao/drawer/drawer.dart';
 import 'package:talao/l10n/l10n.dart';
@@ -58,132 +51,46 @@ class _CredentialsListPageState extends State<CredentialsListPage> {
         }
         return false;
       },
-      child: MultiBlocListener(
-        listeners: [
-          ///Note - Sync listener content with qr code scan listener
-          BlocListener<QRCodeScanCubit, QRCodeScanState>(
-              listener: (context, state) async {
-            if (state.isDeepLink == null) return;
-            if (!state.isDeepLink!) return;
-
-            if (state is QRCodeScanStateHost) {
-              // if (state.promptActive!) return;
-              // context.read<QRCodeScanCubit>().promptDeactivate();
-              var approvedIssuer = Issuer.emptyIssuer();
-
-              var profileCubit = context.read<ProfileCubit>();
-              if (profileCubit.state is ProfileStateDefault) {
-                final isIssuerVerificationSettingTrue =
-                    profileCubit.state.model!.issuerVerificationSetting;
-                if (isIssuerVerificationSettingTrue) {
-                  try {
-                    approvedIssuer = await CheckIssuer(
-                            DioClient(Constants.checkIssuerServerUrl, Dio()),
-                            Constants.checkIssuerServerUrl,
-                            state.uri!)
-                        .isIssuerInApprovedList();
-                  } catch (e) {
-                    if (e is ErrorHandler) {
-                      e.displayError(
-                          context, e, Theme.of(context).colorScheme.error);
-                    }
-                  }
-                }
-              }
-              var acceptHost = await showDialog<bool>(
-                    context: context,
-                    builder: (BuildContext context) {
-                      return ConfirmDialog(
-                        title: l10n.scanPromptHost,
-                        subtitle: (approvedIssuer.did.isEmpty)
-                            ? state.uri!.host
-                            : '${approvedIssuer.organizationInfo.legalName}\n${approvedIssuer.organizationInfo.currentAddress}',
-                        yes: l10n.communicationHostAllow,
-                        no: l10n.communicationHostDeny,
-                        lock: (state.uri!.scheme == 'http') ? true : false,
-                      );
-                    },
-                  ) ??
-                  false;
-
-              if (acceptHost) {
-                context.read<QRCodeScanCubit>().accept(state.uri!, true);
+      child: BasePage(
+        scaffoldKey: scaffoldKey,
+        title: l10n.credentialListTitle,
+        padding: const EdgeInsets.symmetric(
+          vertical: 24.0,
+          horizontal: 16.0,
+        ),
+        drawer: ProfilePage(),
+        titleLeading: IconButton(
+          icon: const Icon(Icons.menu),
+          onPressed: () => scaffoldKey.currentState!.openDrawer(),
+        ),
+        titleTrailing: IconButton(
+            icon: const Icon(Icons.qr_code_scanner),
+            onPressed: () {
+              if (kIsWeb) {
+                showDialog(
+                  context: context,
+                  builder: (BuildContext context) => InfoDialog(
+                    title: l10n.unavailable_feature_title,
+                    subtitle: l10n.unavailable_feature_message,
+                    button: l10n.ok,
+                  ),
+                );
               } else {
-                //await qrController.resumeCamera();
-                context.read<QRCodeScanCubit>().emitWorkingState();
-                ScaffoldMessenger.of(context).showSnackBar(SnackBar(
-                  content: Text(l10n.scanRefuseHost),
-                ));
+                Navigator.of(context).push(QrCodeScanPage.route());
               }
-            }
-            if (state is QRCodeScanStateSuccess) {
-              //   await qrController.stopCamera();
-              ///Note: Push
-              await Navigator.of(context).push(state.route!);
-            }
-            if (state is QRCodeScanStateMessage) {
-              //   await qrController.resumeCamera();
-              final errorHandler = state.message!.errorHandler;
-              if (errorHandler != null) {
-                final color =
-                    state.message!.color ?? Theme.of(context).colorScheme.error;
-                errorHandler.displayError(context, errorHandler, color);
-              } else {
-                ScaffoldMessenger.of(context).showSnackBar(SnackBar(
-                  backgroundColor: state.message!.color,
-                  content: Text(state.message!.message!),
-                ));
-              }
-            }
-            if (state is QRCodeScanStateUnknown) {
-              //   await qrController.resumeCamera();
-              ScaffoldMessenger.of(context).showSnackBar(SnackBar(
-                content: Text(l10n.scanUnsupportedMessage),
-              ));
-            }
-          })
-        ],
-        child: BasePage(
-          scaffoldKey: scaffoldKey,
-          title: l10n.credentialListTitle,
-          padding: const EdgeInsets.symmetric(
-            vertical: 24.0,
-            horizontal: 16.0,
-          ),
-          drawer: ProfilePage(),
-          titleLeading: IconButton(
-            icon: const Icon(Icons.menu),
-            onPressed: () => scaffoldKey.currentState!.openDrawer(),
-          ),
-          titleTrailing: IconButton(
-              icon: const Icon(Icons.qr_code_scanner),
-              onPressed: () {
-                if (kIsWeb) {
-                  showDialog(
-                    context: context,
-                    builder: (BuildContext context) => InfoDialog(
-                      title: l10n.unavailable_feature_title,
-                      subtitle: l10n.unavailable_feature_message,
-                      button: l10n.ok,
-                    ),
-                  );
-                } else {
-                  Navigator.of(context).push(QrCodeScanPage.route());
-                }
-              }),
-          body: BlocBuilder<WalletCubit, WalletState>(
-            builder: (context, state) {
-              var _credentialList = <CredentialModel>[];
-              _credentialList = state.credentials;
-              return Column(
-                children: List.generate(
-                  _credentialList.length,
-                  (index) =>
-                      CredentialsListPageItem(item: _credentialList[index]),
-                ),
-              );
-            },
-          ),
+            }),
+        body: BlocBuilder<WalletCubit, WalletState>(
+          builder: (context, state) {
+            var _credentialList = <CredentialModel>[];
+            _credentialList = state.credentials;
+            return Column(
+              children: List.generate(
+                _credentialList.length,
+                (index) =>
+                    CredentialsListPageItem(item: _credentialList[index]),
+              ),
+            );
+          },
         ),
       ),
     );
